Rename duke files
