+++ build.gradle
@@ -3,22 +3,36 @@ plugins {
     id 'application'
     id 'com.github.johnrengelman.shadow' version '7.1.2'
     id 'org.openjfx.javafxplugin' version '0.0.13'
+    id 'checkstyle'
 }
 
 repositories {
     mavenCentral()
 }
-
-def javaFxVersion = '17.0.7'
+checkstyle {
+    toolVersion = '10.2'
+    configFile = rootProject.file("config/checkstyle/checkstyle.xml")
+}
 
 dependencies {
     testImplementation group: 'org.junit.jupiter', name: 'junit-jupiter-api', version: '5.10.0'
     testRuntimeOnly group: 'org.junit.jupiter', name: 'junit-jupiter-engine', version: '5.10.0'
-    implementation "org.openjfx:javafx-controls:${javaFxVersion}"
-    implementation "org.openjfx:javafx-base:${javaFxVersion}"
-    implementation "org.openjfx:javafx-graphics:${javaFxVersion}"
-}
 
+    String javaFxVersion = '17.0.7'
+
+    implementation group: 'org.openjfx', name: 'javafx-base', version: javaFxVersion, classifier: 'win'
+    implementation group: 'org.openjfx', name: 'javafx-base', version: javaFxVersion, classifier: 'mac'
+    implementation group: 'org.openjfx', name: 'javafx-base', version: javaFxVersion, classifier: 'linux'
+    implementation group: 'org.openjfx', name: 'javafx-controls', version: javaFxVersion, classifier: 'win'
+    implementation group: 'org.openjfx', name: 'javafx-controls', version: javaFxVersion, classifier: 'mac'
+    implementation group: 'org.openjfx', name: 'javafx-controls', version: javaFxVersion, classifier: 'linux'
+    implementation group: 'org.openjfx', name: 'javafx-fxml', version: javaFxVersion, classifier: 'win'
+    implementation group: 'org.openjfx', name: 'javafx-fxml', version: javaFxVersion, classifier: 'mac'
+    implementation group: 'org.openjfx', name: 'javafx-fxml', version: javaFxVersion, classifier: 'linux'
+    implementation group: 'org.openjfx', name: 'javafx-graphics', version: javaFxVersion, classifier: 'win'
+    implementation group: 'org.openjfx', name: 'javafx-graphics', version: javaFxVersion, classifier: 'mac'
+    implementation group: 'org.openjfx', name: 'javafx-graphics', version: javaFxVersion, classifier: 'linux'
+}
 test {
     useJUnitPlatform()
     testLogging {
@@ -27,30 +41,36 @@ test {
         exceptionFormat "full"
         showCauses true
         showStackTraces true
-        showStandardStreams = false
     }
 }
 
 application {
-    mainClass.set("duke.main.Launcher")
+    mainClass.set("lebum.main.Launcher")
 }
 
+// Configure shadowJar
 shadowJar {
-    archiveBaseName = "duke"
+    archiveFileName = 'lebum.jar'
     archiveClassifier = null
+
 }
 
+// Configure JavaFX
 javafx {
-    version = javaFxVersion
-    modules = ['javafx.controls', 'javafx.base', 'javafx.graphics']
+    version = '17.0.7'
+    modules = ['javafx.controls', 'javafx.base', 'javafx.graphics', 'javafx.fxml']
 }
 
+// Configure the run task
 run {
     standardInput = System.in
 }
 
+// Configure the jar manifest
 jar {
     manifest {
-        attributes 'Main-Class': 'duke.main.Launcher'
+        attributes(
+                'Main-Class': 'duke.main.Launcher'
+        )
     }
 }
