--- src/main/java/duke/main/Launcher.java
@@ -1,11 +1,0 @@
-package duke.main;
-import javafx.application.Application;
-
-/**
- * Class that launches the main window
- */
-public class Launcher {
-    public static void main(String[] args) {
-        Application.launch(Main.class, args);
-    }
-}
