+++ src/main/java/lebum/exception/DukeException.java
@@ -1,0 +1,6 @@
+package lebum.exception;
+public class DukeException extends Exception{
+    public DukeException(String message){
+        super(message);
+    }
+}
