+++ src/main/java/lebum/main/Launcher.java
@@ -1,0 +1,11 @@
+package lebum.main;
+import javafx.application.Application;
+
+/**
+ * Class that launches the main window
+ */
+public class Launcher {
+    public static void main(String[] args) {
+        Application.launch(Main.class, args);
+    }
+}
