+++ src/test/META-INF/MANIFEST.MF
@@ -1,3 +1,3 @@
 Manifest-Version: 1.0
-Main-Class: duke.main.Launcher
+Main-Class: lebum.main.Launcher
